(#526) Fix: CopyLimit overflows at 4GB.

+++ Sandboxie/core/dll/file_copy.c
@@ -125,10 +125,10 @@ found_match:
         return mode;
 
     //
-    // if tere is no configuration for this file type/path decide based on the file size
+    // if there is no configuration for this file type/path, then decide based on the file size
     //
     
-    if (File_CopyLimitKb == -1 || file_size < (File_CopyLimitKb * 1024))
+    if (File_CopyLimitKb == -1 || file_size < ((ULONGLONG)File_CopyLimitKb * 1024))
         return FILE_COPY_CONTENT;
 
     //
